+++ haystack/nodes/__init__.py
@@ -22,7 +22,13 @@ from haystack.nodes.file_converter import (
 )
 from haystack.nodes.label_generator import PseudoLabelGenerator
 from haystack.nodes.other import Docs2Answers, JoinDocuments, RouteDocuments, JoinAnswers
-from haystack.nodes.preprocessor import PreProcessor, DocumentMerger, DocumentSplitter, DocumentCleaner, NewPreProcessor
+from haystack.nodes.preprocessor import (
+    PreProcessor,
+    DocumentMerger,
+    DocumentSplitter,
+    DocumentCleaner,
+    DocumentPreProcessor,
+)
 from haystack.nodes.query_classifier import SklearnQueryClassifier, TransformersQueryClassifier
 from haystack.nodes.question_generator import QuestionGenerator
 from haystack.nodes.ranker import BaseRanker, SentenceTransformersRanker
