+++ haystack/nodes/preprocessor/__init__.py
@@ -1,5 +1,5 @@
-from haystack.nodes.preprocessor.preprocessor import PreProcessor
-from haystack.nodes.preprocessor.preprocessor_new import NewPreProcessor
+from haystack.nodes.preprocessor.preprocessor_old import PreProcessor
+from haystack.nodes.preprocessor.preprocessor_new import DocumentPreProcessor
 from haystack.nodes.preprocessor.splitter import DocumentSplitter
 from haystack.nodes.preprocessor.cleaner import DocumentCleaner
 from haystack.nodes.preprocessor.merger import DocumentMerger
